isolate matcher function
